Support for alternate authentication server
When a server gets a request that has a bearer token, if it has a
setting in ego.server.auth.server, it uses that address to form a
request on that server to the /services/admin/authenticate/ endpoint
which verifies that the token is valid. It also returns the array
of string permissions the user associated with the token has.