Move authentication handling into the endpoint router
When an endpoint is defined, it can include whether the endpoint
requires authentication, and if it requires an administrator. When
these are set in the endpoint definition, the router handles the
authentication process, and stores the result in the *Session
object.  If the service requires authentication or an admin user,
the Router will automatically handle 401 and 403 replies.

This lets the handlers be much cleaner and simpler.