Adopt standard secure (443) and insecure (80) port numbers
By default, Ego servers now assume they run on the standard ports for
HTTPS and HTTP servers. If the secure port is used, the server will
also start a redirector service on the insecure port to redirect it
to the secure port (this can be overridden using the --insecure-port
option).