Switch allows local assingment of test value
You can now use something like:

    switch v := mydata[keyvalue].size() {
    ...
    }

and the value 'v' will have the test value in it, available
to any case or default block that needs it. This helps prevent
the need to re-acquire the value by using the expression again
in a non-idempotent fashion.

Additionally, provided a helper function to simplify the most
common case of generating expression code.