Add support for unwrapping interface types
Support for x := foo.(t) to unwrap foo as type 't'. In strict mode, it
must be of type 't' or it is a nil value. Secondary value from the
assignment is a boolean indicating if it succeeded.

Support switch v := foo.(type) which puts the unwrapped value in 'v' and
each case statement is a type check to see what the actual value was.