Move /services/admin/log to a native endpoint by default
You can use the service in the examples folder instead if you wish
by moving it to the lib/services/admin directory. If it is not moved
there (the default) then the native service is used.

This expanded support in the Session object type to track paraameters
and media types to make processing in the handler somewhat simpler.