Implement native handler for /services/admin/authenticate
Like other required admin services, this is now implemented with a
native handler, which can be overridden by putting a handler in the
lib/services/admin folder. An example is in the examples/services
folder.

This fix also adds the redirect needed for logon requests to a server
that is not also the authentication server, and added missing REST
debug logging for response payloads to native admin services.