Start on DSN (data source name) support
A DSN is a way of concealing the database info more thoroughly
from the consumer. When accessing a table, DB credentials are never
sent from the client, or available in plain text on the config
file. They are instead stored in a database (in the same db as
the authentication info).

Access to a table will soon require a DSN to indicate the data
source info used to connect to Postgres to access the table.