Add provider info to DSN
Also, add scaffolding for authorization check of a DSN given a
username. Currently, this just returns true, but will be built
out later. A caller who wants to use a DSN has the responsibility
of using the service.AuthDSN(user, dsname) API to determine if the
DSN is usable; no checking is done on the other APIs.

The auth check includes which action set (read, write, admin) is
to be performed.