fix: if dsn has no schema, use connect identity for schema.
This more closely models how Postgres uses schemas. The best
practice is still to specify a schema in the DSN for the most
comprehensive security, but for privileged users, they may
need to specify arbitrary schemas or want the default schema
of their account information (to the Ego server, not the
account in the connection string).