Add proper encryption to passwords stored in data source names
The password is encrypted using a server-instance-specific token
value based on the ego.server.token.key with additional seasoning.
The string is onlhy decrypted when being used to generate a
connection string to send to the database.

Note that if the server token key is lost or damaged, all DSNS will
be invalid and must be recreated.