Better error message when cipher.Seal() argument is not *string
Also, note that when you seal a string value, the original string
value is set to an empty string. The intent is to remove the
original string value so it can be replaced with the sealed version.
A common pattern is:

    text := "secret value"
    text = cipher.Seal(&text)

This replaces the value of text with the encrypted value.  But if
you do this:

    text := "secret value"
    sealed := cipher.Seal(&text)

The encrypted value is stored in sealed, but the value of text is
also set to an empty string. This is why the value must be passed
as a string pointer to the Seal function.