Use in-memory cache for database-based DSN storage
If an item is created, it is both written to disk and stored in the
in-memory cache. If it is read from database, it is added to the
in-memory cache. If a read of the database can be satisified from
the cache instead, the cache is used. Items in the cache expire after
a minute, so burts of activity will be satisfied without going to disk.
Cache expiration means that if multiple Ego servers are running against
the same database, if a change is made in one server it will be updated
in the cached state of ther servers within about a minute.