Allow command grammar to specify non-zero minimum parameters for varargs
If a command (like "ego dsns delete") supports varying arguments, but must
have at least a minimum number (such as 1), then allow this in the grammar
specification. This requires an additional check during command line parsing
and an update to the help output to show that the argument list isn't
completely optional.