Fix: ensure status field of REST responses always non-zero
This adds code to initialize the Status field of response objects
to http.StatusOK for the default cases of "no error". Previously
the field was zero when there was no error.

Also updated client REST commands to output the JSON if there was
an error and the output format isn't text.