Fix bug in optimizer for constant StoreAlways
The optimizer could not recognize when a StoreAlways operand had
already been converted to a list, so it would make a nested list
which isn't supported. Added a pattern flag that requires that the
operand type be a string, which prevents the pattern from matching
when it is a list intead of a string.