Better implementation of /ui/dsns, using cached HTML template and CSS
The HTML template and the CSS style sheet are both now stored as
assets in the server asset area, so they can be modified without
making additional changes to the code. This significantly simplified
the handler for this request.

Additionally, modified the asset handler to support local loads of
assets (as opposed to those coming from an outside REST call) so
the DSNS list handler could fetch the HTML template file. Added a
new logging class ASSET that reports more fully on when/how an
asset is loaded from disk versus memory, and also reports when the
asset cache is full and older items are discarded.