Add ability to request shorter expiration on a logon
When making a logon request to the server, you can now request an
expiration duration in the credentials payload. If this value is
less than (or equal to) the default server expiration, the resulting
token will have the shorter expiration.

For example, this would allow a token created for a unit test to
be specified as expiring within 10m so it is not valid for subsequent
tests after that time.

As an extension, the ego command line version of this (the --expiration
option) allows the expiration to be specified as days, such as "30d".
This is converted to hours before passing to the server.