Fix: go routine can call bytecode or native functions
The Go processor was passing the function to invoke as a name that
was retrieved from the runtime symbol table. The change is to allow
the code to be compiled normally so the Call operation either uses
a literal function pushed on the stack, or a function invocation
using a named item. The Go processor replaces the Call operation
with a Go operation, which can transcribe the arguments and the
function reference to a new temporary bytecode stream launched as
the go routine.