Clean up inaccurate/unhelpful argcount in reflect.Reflection declaration
In the reflect.Reflection{} structure for a function, the function
declaration object now more accurately describges the argument counts.
If the function is variaadic, the Argcount array sets the second (max)
argument count value to -1 to indicate varying. Additionally, the
first (min) argument count accurately reflects the minimum number of
arguments that must be specified for the function.