Add --service and --asset selector to server cache show command
You can specify --service or --asset to limit the output to only
that specific cache class. If a specific class is requested, the
"Class" column is omitted from text output. If the output format
is json, then using the selector will show only cached infor of
that class in the "items" array.

Note this is entirely a client function; the server returns the
full list of cached items regardless of this selection.