Automatically create minimial lib/ directory if not found
When running a program or starting a server, Ego will check to
verify that the designated EGO_PATH location exists. IF it does
not, then it will create a minimally populated lib direcgory in
the EGO_PATH location.

This automatic initialization can be suppressed using the private
global option --no-lib-init, or setting EGO_NO_LIB_INIT environment
variable to "true".

Note that anytime a change is made to the lib tree for the Ego
project, the zipgo command (https://github.com/tucats/zipgo)
should be run to re-generate the internal zip-file represenation
of the library tree.