Change _server_instance global variable to _instance
This value is a unique UUID for any execution of the Ego language
processor, whether it is in server mode or not. So the name should
reflect that it is not server-specific. This value is always set
to a new UUID on initialization, but may be overridden when the
UUID is supplied as a command-line option.