Ensure package data is always readonly when copied from another table
This ensures that the shared data between the source and destination
symbol tables is readonly. Also removed dead code for Copy() method,
which did a shallow clone of a table. Also updated the swarm.ego
example to dynamicaally acquire the server hostname instead of it
being hard-coded into the example.