Fix thread safety issue with Ego REST server response struct
The struct was stored in the symbol table for the cached service
instance, so it was inherently not thread-safe, and multiple
concurrent uses of the same service created corrupted response
body structures.

The fix is to add a new symbol scope unique to the execution of
the response handler that holds the $response structure.

This update also restructures the defs package contents to more
easily identify where a given definition item can be found in the
source files.