Add static redirects to the Ego server
If the file "redirects.json" is found in the lib directory of
the Ego path, it contains a JSON dictionary of static redirect
operations added to the routing table when the server starts.
This can be used to create any arbitary redirects from URL paths
relative to the Ego server itself.

Default redirects are provided in the sample file that support
redirecting request to the /docs path to the online HTML doc
files.