Server can be configured to run service requests in a subprocess
The server by default runs service reequests in a thread on the main
server process. You can configure the server instead to run each
service on a single subprocess. This provides better isolation for
services competing for resources, and allows for a quota to be
set for how many service processses can exist at one time.