Complete rewrite of the fmt.Sscanf() scanner
The previous version was attempting to use the tokenizer to assist
in recognizing scannable objects in the input string, but this was
not behaving correctly to Go standards regarding blank space and
width specifications. This version discards the use of the tokenizer
and scans through the string byte-by-byte. It also more-correctly
handles white space in the format and the data strings.