If extensions enable, allow a var type of "type"
This creates a variable that holds a type definition. This can
be assigned to a type ( foo := int ) or a type can be parsed and
assigned to it (fmt.Sscanf("int", "%t", &foo)). The resulting
variable can then be used as a type itself (var bar foo).

This is only supported when language extensions are enabled.