Support zip archive of log files that age out
By default, when the server restarts or the log file rolls over at
midnight, older logs (exceeding the retain count) are deleted. When
the --archive-log global option is used, or the ego.log.archive
config setting has a non-empty value, instead of deleting the old
log files they are added to this zip file for archival purposes.