New config ego.runtime.client.timeout sets timeout duration for REST calls
Also, if a call is taking more than a few seconds, logging messages are sent
to the console indicating that the code is still waiting on the REST call to
complete. The timeout can be set to "none" or "0" and there will be no
timeout.

Also, make "status" the default subverb of ego server command.