Improved handling of archive checksums for zip, lang
Previously, the checksum was stored in a separate file for the
generated files. However, there's no reason not to store them
as comments at the top of the output file. This modifies both
tools to use the existing source file (if present) to check
for a matching digest checksum value.