Add ego.runtime.insecure.client config option
When set, this disables TLS certificate checking in the client.  Note
that this _cannot_ be set by a command-line option; it must be modified
in the ego.json file directly. This is meant to reduce the chance that
this would be set accidentally, as it is _only_ suitable for use as a
debugging tool and should _never_ be used in a production deployment.