Convert single JSON config file to per-profile JSON files
The files are named with a ".profile" extension in the user's home
directory under the .ego subdirectory. The intent of this is to make
it easier to share a profile, simply by moving the file from one
location to another -- difficult when the profiles were stored as a
single array of structures.

With this change, the next time ego is run, it will automatically
convert the profile data to discrete files. You can track the
actions of the profile loader by enabling the APP logger at
startup with the EGO_DEFAULT_LOGGING environment variable, i.e.

   export EGO_DEFAULT_LOGGING=APP