Crude localization of help text, better English help text
The help text was reformatted to be more consistent in an 80-column
terminal window. Additionally, numerious formatting, spelling, and
grammar errors were corrected.

A super-simplified version was created in French and Spanish to test
the selection of an available help file for those languages.  Note
that these are NOT functional translations, and only include the
initial "introduction" help entry.