Fix bad error handling of signed constants in table filters
There was an error in the parsing of a filter term that did not
allow for a signed constant value.

This also exposed a general weakness in the generator and parsing
logic that did not cleanly propogate errors well, which could
result in bogus SQL being generated. This fix adds error return
codes to the parsers and generators for filters and SQL to better
handle reporting the actual error(s) encountered in a query
processing phase.