More complete column metadata for abstract row reads
The columns array now contains objects that have the name, sql type,
and length (if bounded) for each column. If the value is unbounded
or the length is not supported by the drive, this returns -1 as the
length.

This format allows much more control over interpreting the json data
for each column, at the expense of not being able to use json to
directly unmarshal the result row into a structure.