Improved time.Duration formatter
This knows about days, and formats durations of greater than 24 hours
accordingly. This is different than the standard Go package. Also,
allow a nonstandard paramter to the String() method for the duration
that, when set to true, adds extra spaces to the output for better
readability. This is used by the /services/up service.