Add Ego call stack to panic() function
If an Ego program includes a panic() operation, then before the
panic is processed, an Ego call stack is dumped to the console.
This is true whether the panic() becomes a "hard panic" or just
stops execution of the Ego program. In the case of the panic()
the call stack symbol table names are not included in the
output.

Also, this fix corrects a number of tautological errors in the
math tests.