Improved security for http/https redirects
If an Ego endpoint does not require authentication or permissions,
then an automatic redirect is allowed. Otherwise, the attempt to use
the HTTP method on a secured server results in an error indicating
that the HTTPS method must be used for this endpoint.

Also improved messages to the log, such that redirected sessions
generate a new session sequence number and use that in the log.