Make data type of "type" a language extension
When language extensions are turned off,
1. You cannot define a variable of type "type"
2. The reflect.Type() function is disabled
3. You cannot assign a value of type "type" to a variable

In particular, this helps trap the common programming error of
specifying

   a := []string

When extensions are enabled, this is perfectly okay as it assigins
a type of "[]string" to the variable a. But without extensions, this
is an invalid statement, and probably means the programmer omitted
the initialzation, i.e. it was meant to be

   a := []string{}