Add Find() method to tables.Table objects
This method calls a closure function that indicates which row(s)
are to be considered found in the contents of a table. For example,

	rows := t.Find( func(columns []string ) bool {
		age, _ := strconv.Atoi(columns[1])
		return age > 50
	})

The closure is called with each row in the table as an array of
strings, where each element is a column in the given row. The
closure makes any determination it wishes to determie if the
row should be considered "found" and returns true or false.
The Find() method returns an integer array of all the found
rows.