Fix Find() method for table.Tables to use variadic arguments
Instead of the Find closure function always being an array, the
function is now declared as variadic list of string values. This
allows the function to be defined as taking a variadic string array
or having explicit arguments for each column (though in this case
the number of elements in the argument list must match the number
of columns in a row).

This makes for a more readable and self-documenting find closure
function.