Fix struct initialization to convert values to struct field types
If you define a struct like this:

   a := struct{
      age  int
      name string
   }{ age:64.4, name:"Tom" }

the code would incorrectly convert the type of field "age" from an
int to a float when dynamic typing was enabled. Because there is
an explicit type, the value of the initializer must be converted
to the type of the field. For static typing, this declaration would
result in an error instead of converting the initializer.