Support pointer objects in serialization
When a bytecode object is serialized, any pointers (types or other
bytecode objects) are not serialized in place, but instead added to
a "pointers" member is the JSON object. Any references to the same
type or bytecode is a referenced to a value in the "pointers" list.
This would prevent a reconstitution of the serialized object from
trying to create distinct instances of the same value being
pointed to.