Fix bug with scope handling accross function boundaries
The code was incorrectly allowing a non-literal, non-closure function
to access symbols outside its scope that were not global values.  That
is, a nested function could always access the symbols that were in the
enclosing function.

This fix creates a "scope barrier" when a function is defined that
prevents the symbols from being accessed incorrectly. However, when
running in "test" mode, this is turned off, so blocks and functions
used in a test stream will work when combined with any other test
streams.