Add @symbols [label] directive
This dumps the symbol table tree for the current scope and thread
to the console. If the optional label is present, it is put in
the header of the output, otherwise the current executing context
name (typically the function signature) is used.

This is serialized, so multiple threads using the @symbols
directive will have their output generated as a unit, not
interleaved with other symbol output, to enhance readability
of the output.

Note that _other_ output from the thread/go routine not
serialized.