Optimize symbol table pop to ignore tables that are unmodified
As each table is popped, if it is not a modified clone for a
package, it is skipped for scanning for updates to package
parents. This means that a cloned table that doesn't have a
global variable that's been updated is not processed, which
speeds up the call frame pop operation.