Fix issued with embedded types not initializing correctly.
When a struct type is embedded in another struct, the initializer
for a list of values was not correctly processing the embedded type
(it was being assigned to the first field in the type, instead of
being applied to all fields provided in the type).

Added a unit test to validate this behavior. Also removed an unwanted
debugging message in one of the unit tests.