Fix issue with explicit import of "os" when autoimports disabled
Turning off autoimports cases the runtiem initializer to only
create a minimum instance of the os package to support Exit().
However, the code was not adding the other builtin-package
members when an explicit import of "os" was done by the user.

This detects this case and complete the runtime initialization
of the "os" package as well as loading any source elements.