Fix bugs in explicit import when auto-import disabled
The "os" package was not correctly importing when auto-imports were
disabled. This was in part because it was already partially constructed
when in no-auto-import mode to support os.Exit() but was not being
correctly added to by a subsequent import. Additionally, a subsequent
import was incorrectly adding all the builtins to all other builtin
packages.