Code hygiene cleanups
The golangci-lint tool was not running because the configuration of
the tool was specifying the malign lint test, which is deprecated.
This was causing the entire tool to abort, so no linting has been
done for at least a few months. I've corrected this error in my
IDE configuration, and now cleaned up the complaints.