Update io.ReadDir() to return modifiation date as *time.Time value
Previously this was returned as the modification date formatted as
a string value. With the better encapsulation of the Ego time.Time
value, the structures returned by io.ReadDir now contain better
user-managable data. The io.DirList() function was modified to
format the time.Time value correctly.

This change also includes updates to the type initializers for all
builtin types to include a serialization mutex so multiple threads
cannot initialize the type simultaneously.