Fix optimization for multiple attempt to initialize package types
A previous optimization had to be removed, this try should be more
correct. This update prevents the unnecessary recompilation and/or
construction of the type values which are anchored in the global
variables for the runtime packages, but the complete package
definition is recreated and reapplied to the current symbol table.

This prevents multiple type definition operations, but does register
the package for each unique service instance or go routine thread.