Enable ephemeral cache for tokens, user identities
When a userid must be looked up in the user database, or a token
must be authenticated, the system first looks in the ephermeral
in-memory cache. If found, the value in the cache is used.

Once a minute, the cache is flushed of any items in the cache
older than one minute. This means that repeated server requests
will be satisfied from the cache if possible. However, changing
a value in the underlying database (or a token expring) will be
acted on within sixty seconds as the cache item will be discarded
and need to be reloaded/re-evaluated.