Fix detection when bytecode is no longer sealed or optimized
Add logic to indicate when the bytecode array is no longer sealed
(that is, truncated to the actual number of bytecodes in the array)
or if a change has been made that invalidates the optimized state
of the bytecode (for example, adding or deleting instructions).

This also prevents the optimizer from being run more than once on
an optimized bytecode fragment.