Add Ego linenumber profiling controlled by --profiling option
If this option is used on the run command, Ego dumps a table showing
how many times each line of the code was executed.

This can be selectively enabled or disabled using directives:

   @profile start
   @profile stop

And a report can be generated using

   @profile report

If the report option is never given and profile data was collected,
it is printed when Ego exits.