Fix issues with line numbers in debugger
The debugger was failing to access the correct line source when
calling functions within packages. This fix causes the compiler to
emit alternate AtLine instructions when the debugger is present
that include the source text in the opcode argument. The debugger
will use this if present in the AtLine instructions that send
control back to the debugger.

If the compiler is not told that debugging is active, it will not
emit the source text in the AtLine instructions, making the normal
execution mode no less efficient than before.