Fix show source in debugger not being able to access package source
The show source command relies on access to the tokenizer used to
parse the source file. When calling into a function not in the
current source file (such as a package function written in Ego),
the compiler how includes a pointer to the active tokenizer for
each source file in the call frame. This allows the debugger to
be able to list the active source file instead of just the source
file run by the command line.