Native function bindings for simple package functions
Updated many of the strings package functions that used wrapper
functions to call native Go versions to directly call them from
the Call bytecode function using reflection.

Supported functions can return a single value, a list of values,
or an array of string values.

This capability will be extended to other packages in the future.