Complete conversion of time package to native functions
All previous implementations of the time package have been converted
to native passthrough calls to the Go time package directly, except
for the (t time.Duration) String() function, which is still impelmented
as a shim, so it can support the optional extended formtting for
durations.

Updated the unit tests according, and deleted the previous shim
functions no longer needed.