Allow assign shortcut in ego config command
If you issue `ego config foo.bar` the value of foo.bar is displayed.
This is because the default config command is `show` in this case.

This update means if you issue `ego config foo.bar=1` it willa ssume
the command is `set` and will set the value.

Additionally, updated errors for argument types including the parameter
number when possible.