Add global option '--cpus n' which sets max number of cpus to schedule go routine upon
For a Go routine that is compute-only, it turns out to (as of Go 1.23.2) be faster to
run with _less_ than the total number of cores. So for an Apple M4 Pro, the default is
14 cores. Running a test that spawns 10,000 threads that each compute the factors of an
integer that uses all the cores runs nearly twice as slowly as running the same program
(with the same number of Go routines) when limiting them to only four cores.  As far
as I can tell so far, this is a quirk of the Go schedule and not an Ego issue, but
research continues...