Improved duration parsing when extensions enabled
The previous duration parser was extended to allow days "5d" as the
duration string, but this could not be mixed with other units. The
new parser allows durations to include days with other units, such
as "3d 1h" which is three days, one hour in duration.

This fix also includes updates to comments in the code for enhanced
clarity to the reader.