fix: correct error where --help did not exit after displaying help
The tokens already parsed were still used to attempt to execute the
action. So "ego server status -h" would display the help and then
still attempt to display the server status.

Also, fixed a potential fatal panic for User and DSN objects that
are stored in the internal cache; if the cache object was incorrectly
typed it would have paniced. Added Go unit test to validate that the
type is now being checked correctly.