fix: process sandboxing when marshalling file names to native functions
If the native function has a string argument that is a filename, a new
attribute (Sandboxed) indicates this name is a filename paarameter and
the sandboxing filter should be applied. When sandboxing is enabled, it
forces all file names to be relative to a predetermined path (the sandbox)
which isolates file I/O to a given directory tree.