fix: defer function not working when calling a reciever method
The defer statement was not capturing that a receiver was specfied
in the call. The fix captures the reciver stack created for the defer
statement and makes the same stack available when the defered statement
are run at function exit.

Also, cleaned up formatting of numerous native and internal object types,
so you can print an *os.File or time.Duration value and get a meaningful
result.