fix: better validation of login parameters
The order of REST logging and paraemter handling made for a difficult
read. Now, the duration is validated and the login server validated
first. If the username or password must be prompted for, that happens
next. Then the rest log shows the results of the actual login operation.

Additionally, cleaned up the error message for a login host error;
the low-level error was not helpful in most cases and hard to read for
a non-admin user.