feat: --new-token on server start or restart will regenerate server token
Using this causes the server initialization to randomly generate a new
token value. This is particularly useful on a server restart as it will
automatically invalidate all tokens generated by the server, forcing all
users to log in again.

Note that using this option in a server start command does not cause it
to be retained for future restarts; it is only used once during the initial
server restart unless explicilty put on subsequent server restart commands.

Also fixed a bug where removing an outboard encrypted configuration setting
(like ego.server.token.key) was not correctly deleting the outboard file
in the config directory.