fix: incorrect symbol scope for defer function literal
When a function literal is used in a defer, the body of the literal
should have access to the outer symbol scope _at the time of execution_
of the defer operation.

The fix was to correct an issue with the symbol scope for the execution
of the defer being marked as a scope boundary. Additionally, the
system now correctly captures the symbol table at the time of the
defer to ensure that subsequent scope changes do not affect how the
defer function sees the symbol scope.

Added an integration test to validate.