fix: fix issue with http response variable scope revealed by previous defer fix
The scope boundary flag was being set incorrectly in the past, and this was
correted as part of fixing how the defer closure function can access local
symbols. This broke the use of the special $response variable stored in a
special scope by the REST service handler which made it inaccessible. This
fixes addresses that by allowing access to this value to bypass scope
boundaries.