Fix: errors in HTTP built-in support for Response methods
A number of the methods incorrectly referenced values outside of
scope for the method (revealed by a recent fix that corrected a
significant symbol scope boundary error). This corrects the code,
and adds documentation to the http import file to explain the
methods.

Also updated the helper tool for messages to print a more useful
and complete message.