fix: better error handling when using SQL for resoruce object mgmt
A number of cases, such as bad filter construction, resulted in
either incorrect errors, panics, or no corrective behaviro at all.
This now causes dependent or chained operations like filter
specification to store any error recorded in the resource handle
and percolate that to the caller of the resource operation.