fix: catch type coercion errors rather than just assume zero value
The coerce and normalize operations were silently discarding errors
where conversions were not possible, and isntead just returning a
zero or nil value.

This change modifies the coerce and normalize operations to percolate
an error code which will be reported when automatic conversion is not
possible.

This also revealed a number of cases where the check for coercsion
errors indicated incorrect support for some operations (array add
and subtract, for example) as well as incorrect handling of interface
arguemnts in some cases.