feat: add ego.runtime.precision.errors to control if precision loss is an error
The default is true; if a conversion causes a signfiicant loss of precision,
then an error is returned. When set to false, the cast may result in trunctation
or other data loss.

Updated the integration test for cataching loss-of-precision errors to skip when
the configruration item is set to false.