feat: better handling of rune values, and extended support for rune arrays
Support 'X' token as a 32-bit integer rune value, suitable for use where
a rune is needed (strings.Builder WriteRune(), for example).

It is an error if the rune string has more than one rune value in it, unless
compiler extensions are enabled, in which case the rune string becomes an
array of int32 values. I.e., 'ABC' is []int{65, 66, 67}