feat: allow Go-native types to use custom formatter function
If a type for a Go-native type (such as strings.Builder) is defined
with a formatter function, that is called when asked to format the
object. Note this is different than the String() method, which formats
the contents of the object as a string. For example, the formatter
for a strings.Builder object will report the length, capacity, and
current buffer contents when asked to print the builder object.