feat: compiler detection of unused/ineffectual variable create or assign
The compiler will attempt to detect when the code creates a variable (by
assignment or with a var statement) and never reads this value. When this
is true, an error is reported indicating the end of the block where the
error was found.

This can be turned off with the ego.compiler.unused.var flag; the default is
true which means it is an error to have an unused value. If false, this error
is ignored. The default value is "true".

This fix includes numerous changes to the integration tests that were createing
values that were never used.