fix: do not re-read config from disk when changing default config
When the --profile option is used to specify a specific profile, the
code was rereading all profile data into memory before selecting the
new profile. This change causes the code to realize the base profile
has already been read, and only read in the outboard config items
like logon token which are not read until a profile is active to
reduce the in-memory footprint of secure items.