feat: allow logs to be formatted as JSON
ego.console.log.format can be "text", "json", or "indented" and affects the format
of log messages sent to the console or a log file.

If log messages are localized text with "log." prefix, they will be converted and
the argument map appaers in the JSON. Otherwise, the JSON contains the text of the
log message as generated for text output.

Conversion of log messages to localized format will support language-independent
log files. A future change will allow the ego log command to translate JSON log
files back to text in the current localization.