feat: enhanced formatting options for localization strings
The {{text}} substitution string can now have additional formatting
values added.  The items are separate by a vertical bar "|" within
the brace set. The first item must always be the key value to be
retrieved (same as the value always was used for). Subsequent items
in the brace set define additional formatting rules.

  %format      Any Go formatting string may be used in place of just
               formatting the item using %v which is the default.

  list         If the item is a list or map, format it with commas
               between items for readability.

  empty "t"    If the value to be formatted is empty or the zero
               value for it's type, use the text "t" in place of
               the empty value.

  label "t"    If the value is not the zero value for it's type,
               then prefix the value with the text "t" to act as
               a label. If the value is zero/empty then the label
               is not printed.

For example,

   Permissions: {{perms|list|empty none}}

If the "perms" value in the log mapping entry is a list, it will be
printed with comma-separated value. If the list ie empty, the string
"none" is printed instead.